requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.22.0
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
banks==2.2.0
bcrypt==5.0.0
billiard==4.2.4
blinker==1.9.0
build==1.3.0
cachetools==6.2.4
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cohere==5.20.0
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.1
fastapi==0.125.0
fastavro==1.12.1
filelock==3.20.1
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
gitpython==3.1.45
google-auth==2.45.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
griffe==1.15.0
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==1.2.3
humanfriendly==10.0
idna==3.11
importlib-metadata==8.7.0
importlib-resources==6.5.2
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
kubernetes==34.1.0
llama-cloud==0.1.45
llama-cloud-services==0.6.88
llama-index-core==0.14.10
llama-index-instrumentation==0.4.2
llama-index-workflows==2.11.5
llama-parse==0.6.88
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy-extensions==1.1.0
narwhals==2.14.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.13.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pip==24.0
platformdirs==4.5.1
posthog==5.4.0
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==5.9.8
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.3
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydeck==0.9.1
pygments==2.19.2
pymupdf==1.26.7
pypdf==6.4.2
pypika==0.48.9
pyproject-hooks==1.2.0
pyreadline3==3.5.4
pymilvus==2.5.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pyyaml==6.0.3
-e file:///C:/Users/dat/Projects/RAG_Native
rank-bm25==0.2.2
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.45
starlette==0.50.0
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.38.0
vine==5.1.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.0.1
yarl==1.22.0
zipp==3.23.0
